feat: rename ResolveUserContextAction and UpdateProfile to ResolveUserContextService and UpdateProfileService for consistency
