+++ app-modules/message/src/Actions/NewMessage.php
@@ -7,8 +7,8 @@ namespace He4rt\Message\Actions;
 use He4rt\Character\Entities\CharacterEntity;
 use He4rt\Message\DTO\NewMessageDTO;
 use He4rt\Provider\DTO\ResolveUserProviderDTO;
-use He4rt\User\Actions\ResolveUserContextAction;
 use He4rt\User\Models\User;
+use He4rt\User\Services\ResolveUserContextService;
 use Illuminate\Support\Facades\Log;
 use Throwable;
 
@@ -29,7 +29,7 @@ final readonly class NewMessage
                 'username' => $messageDTO->providerUsername,
             ]);
 
-            $userContext = resolve(ResolveUserContextAction::class)->handle($userDto);
+            $userContext = resolve(ResolveUserContextService::class)->handle($userDto);
 
             $userContext->character->refresh();
